requirements.txt
bleach==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==1.11.29
djangorestframework==3.9.4
docutils==0.17.1
idna==2.10
packaging==20.9
pkginfo==1.7.1
Pygments==2.5.2
pyparsing==2.4.7
pytz==2021.1
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
tqdm==4.61.2
twine==1.15.0
urllib3==1.26.6
webencodings==0.5.1
api-auto-doc==0.1.0